+++ client/src/pages/Manage.tsx
@@ -1,20 +1,6 @@
-import { Link as RouterLink } from "react-router-dom";
-import { Link as MUILink, Typography, Box } from "@mui/material";
+import { Typography, Box } from "@mui/material";
 import Grid from "@mui/material/Grid2";
-
-const buttonStyle = {
-  display: "inline-block",
-  padding: "8px 16px",
-  color: "white",
-  backgroundColor: "primary.main",
-  borderRadius: "4px",
-  textDecoration: "none",
-  width: "100%",
-  textAlign: "center",
-  "&:hover": {
-    backgroundColor: "primary.dark",
-  },
-};
+import BtnLink from "../components/BtnLink";
 
 export default function Manage() {
   return (
@@ -33,29 +19,13 @@ export default function Manage() {
         <Grid size={3}></Grid>
         <Grid container spacing={2} size={6}>
           <Grid size={12}>
-            <MUILink
-              component={RouterLink}
-              to="/manage/competitions"
-              sx={buttonStyle}
-            >
-              Gérer compétitions
-            </MUILink>
+            <BtnLink to="/manage/competitions" content="Gérer compétitions" />
           </Grid>
           <Grid size={12}>
-            <MUILink component={RouterLink} to="/manage/teams" sx={buttonStyle}>
-              Gérer équipes
-            </MUILink>
+            <BtnLink to="/manage/teams" content="Gérer équipes" />
           </Grid>
           <Grid size={12}>
-            <MUILink
-              component={RouterLink}
-              to="/manage/juries"
-              sx={buttonStyle}
-            >
-              <Typography variant="body1" component="span">
-                Gérer jurys
-              </Typography>
-            </MUILink>
+            <BtnLink to="/manage/juries" content="Gérer jurys" />
           </Grid>
         </Grid>
         <Grid size={3}></Grid>
